kernel: hil: digest: only one client
Dividing client traits to match the downcall traits leads to challenges.
The main issue is it is difficult for implementors of split traits to
keep track of multiple clients.

Our first attempt to help this was to define another trait `Client`
which groups all traits. This leads to many set_client() calls which are
just empty.

The result is that we lose static guarantees because we can call a valid
set_client which is not the set_client that the capsule expects. This
leads to a runtime failure which should be a compile time failure.

This is one example of addressing this, by removing the split traits on
just the up-call (client) side. Now there is only one client type to
worry about.